requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
cerebras-cloud-sdk
numpy==1.26.3
faiss-cpu>=1.8.0
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers>=2.3.0
torch>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.0
